requirements.txt
Flask==3.1.1
flask-cors==6.0.0
openai==1.82.0
pyperclip==1.9.0
python-dotenv==1.1.0
